Move token-related code to token.py
