+++ calc/__init__.py
@@ -22,4 +22,5 @@ __author__ = 'Reilly Tucker Siemens, Alex LordThorsen'
 __email__ = 'reilly@tuckersiemens.com, alexlordthorsen@gmail.com'
 __version__ = '0.1.0'
 
-from .calc import INTEGER, EOF, PLUS, Token, Calc, CalcError
+from .calc import Calc, CalcError
+from .token import INTEGER, EOF, PLUS, Token
